service.copernicus/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
pystac-client==0.7.5
rasterio==1.3.9
shapely==2.0.2
geojson==3.1.0
httpx==0.26.0
python-multipart==0.0.6
numpy<2.0.0